+++ service/stream.go
@@ -70,3 +70,12 @@ func (s *StreamFactory) GetRecentFinalBlock() (uint64, error) {
 
 	return finalBlockNum, err
 }
+
+func (s *StreamFactory) GetHeadBlock() (uint64, error) {
+	headNum, _, _, _, err := s.hub.HeadInfo()
+	if err != nil {
+		return 0, err
+	}
+
+	return headNum, nil
+}