+++ tui2/replaylog/file.go
@@ -3,6 +3,7 @@ package replaylog
 import (
 	"fmt"
 	"io"
+	"log"
 	"os"
 
 	pbsubstreamsrpc "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v2"
@@ -10,9 +11,10 @@ import (
 	"github.com/streamingfast/dbin"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/streamingfast/substreams/tui2/stream"
 	"google.golang.org/protobuf/proto"
 	anypb "google.golang.org/protobuf/types/known/anypb"
+
+	"github.com/streamingfast/substreams/tui2/stream"
 )
 
 type File struct {
@@ -81,6 +83,11 @@ func (f *File) ReadReplay() (out stream.ReplayBundle, err error) {
 			if err == io.EOF {
 				break
 			}
+			if err == io.ErrUnexpectedEOF {
+				log.Println("Unexpected EOF reading replay file, ignoring it.")
+				break
+			}
+
 			return nil, fmt.Errorf("reading replay file: %w", err)
 		}